+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './ActionButton'
 export * from './AddMembersPaper'
 export * from './App'
 export * from './AppRouter'
