+++ vite.config.ts
@@ -7,12 +7,6 @@ import wasm from 'vite-plugin-wasm'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react(), tsconfigPaths(), wasm()],
-  optimizeDeps: {
-    exclude: [
-      'chunk-YSDLPTTY.js',
-      'chunk-XBHG3ID7.js'
-    ]
-  },
   esbuild: {
     supported: {
       'top-level-await': true, //browsers can handle top-level-await features
